third_party/WebKit/LayoutTests/fast/canvas-api/gradient-addColorStop-with-invalid-offset.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function() {

    var gradient = document.createElement('canvas').getContext('2d').createLinearGradient(0, 0, 150, 0);

    assert_throws("INDEX_SIZE_ERR", function() {
      gradient.addColorStop(-1, '#000');
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      gradient.addColorStop(2, '#000');
    });
    assert_throws(new TypeError(), function() {
      gradient.addColorStop(Infinity, '#000');
    });
    assert_throws(new TypeError(), function() {
      gradient.addColorStop(-Infinity, '#000');
    });
    assert_throws(new TypeError(), function() {
      gradient.addColorStop(NaN, '#000');
    });
}, 'This test checks invalid offsets on gradients.');
</script>
